+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Changelog
 
+## 1.7.1
+- Refactor file upload handling in Crudable trait to improve storage configuration management
+
 ## 1.7.0
 - Changed `selectedItemId` to `selectedId` and updated necessary components
-- Updated `cancel` method to reset forms and dispatch 1close-modal` event
+- Updated `cancel` method to reset forms and dispatch `close-modal` event
 
 ## 1.6.0
 - Update button variation components